Add option to change the tag position
Added a option to change the tag position of a new tag as default behaviour displays the tag first before search results. Passing `:tag-position="bottom"` will revert this behaviour and will display search results before the new tag which will allow the user to press enter to select a search result if it already exists.

If it's a completely new option then there will be no search results remaining and so the new tag will be highlighted allowing a user to press enter to add the new tag.