Rough color extraction
Not all shades of grey have been extracted. It feels like there
are too many variants of only very minor different shades of grey.

Hower, this makes sense when looking at the component visually. It
should in any case be determined wether to extract all the greys or
wether it may be possible to consolidate some.