refs #420: Tried (and failed) implementing the feature
This is trickier than expected. It seems rather straight-forward
to expose two text extractions from the <style>-Block of a vue
component, one with the appropriate style loaders, one as a
raw text extraction.

In a perfect world, these two extractions could be done simulataneously
and result in two exported chunks. However, my research led me to
the multi-loader which is supposed to be able to do exactly that but
in practice yielded error messages I fail to understand.

Hopefully, there is a better solution to tackling this.

At any rate, this currently exports .js and .scss and should
thus NOT be merged with upstream without changes eventhough
things should otherwise work as expected.